Parameters for Predict and Numerical Optimization in Kalman Filter

Setup in the predict and numerical optimization in the Kalman filter algorithm.

car_control(fty=1, n.ahead=10, trace=FALSE, ari=TRUE, vri=FALSE, vr=0, pfi="MAPS",
ccv="CTES", lpv=TRUE, scc=TRUE,  nit=40, opm=1, rgm=1, req=0.5, con=1.0e-5, rpe=1.0, 
ivl=1.0e-2, fac=1.0e1, stl=1.0e-5, sml=1.0e2, gtl=1.0e5, kst=TRUE, fct=TRUE)

Arguments

fty
fty=1 forecast past the end. fty=2 forecast last L-steps. fty=3 forecast last L-steps updated (filtering)types. See also fct. If fct=TRUE, all time series is used to fit the model for fty=1 or 3. If fty=2, only the first (length of time - n.ahead) is used to fit the model. Thus, only the first (length of time - n.head) prediction values are the same for fty=1 or 3 vs fty=2. See also pre2, prv2.

n.ahead
number of steps ahead at which to predict.

trace
a logical value triggering printout of information during the fitting process, and major results for the fitted model.

ari
ari=TRUE: parameter starting values. ari=FALSE: they are taken as zero. This is obsolete.

vri
vri=FALSE, observation noise not included in the model. vri=TRUE, observation noise included

vr
0.5, initial value of observation noise ratio: only if vri=TRUE

pfi
always use the option pfi="MAPS".

ccv
ccv="CTES" for constant term estimation. ccv="MNCT" if mean correction, ccv=NULL if omitted.

lpv
lpv=TRUE always use this option.

scc
scc=TRUE always use this option.

nit
number of iteations.

opm
opm=1 always use this.

rgm
rgm=1 always use this.

req
root equality switch value.

con
convergence criterion.

rpe
relative size of parameter perturbations.

ivl
initial value of step size constraint parameter.

fac
step size constraint modification parameter. This value may be setup to fac=5 for better convergency.

stl
typical smallest step size parameter.

sml
typical small step size parametrr.

gtl
typical greatest step size parameter.

kst
kst=TRUE to save estimated states.

fct
fct=TRUE to use all time series to fit the model.

Details

Objects returned by this function specify predict and numerical optimization parameters of the Kalman filter algorithms implemented in car, (via the ctrl argument).

Value

An object of class car_control, a list.

See also